requirements.txt
streamlit
torch
torchvision
numpy
pillow
opencv-python-headless
scikit-learn
matplotlib
